Files changed:
contact.html
faq.html
index.html
mituri.html
psihoterapia.html
script.js
style.css
style.css.map
style.scss
Changes:

HTML files:

All HTML files (contact.html, faq.html, index.html, mituri.html, and psihoterapia.html) have the same changes:
A new <link> tag was added to the <head> section, pointing to a favicon at /images/logo cristina anghel.svg.
A new <button> element was added to the <body> section, with an id of myBtn and a title of "Go to top".
JavaScript file:

script.js:
A new function scrollFunction() was added, which checks if the user has scrolled down more than 20 pixels from the top of the document. If so, it displays the "Go to top" button.
A new function topFunction() was added, which scrolls the user back to the top of the document when the "Go to top" button is clicked.
CSS files:

style.css:
A new CSS rule was added for the #myBtn element, styling the "Go to top" button.
The :hover pseudo-class was added to change the background color of the button on hover.
style.css.map:
The file was updated to reflect the changes made to style.css.
style.scss:
The same CSS rules as in style.css were added, but in SCSS syntax.
In summary, these changes add a "Go to top" button to all HTML pages, which becomes visible when the user scrolls down more than 20 pixels from the top of the document. The button is styled using CSS, and the JavaScript code handles the scrolling behavior.